# materials/api-docs.json
{"swagger":"2.0","info":{"description":"Describing the purpose of the API","version":"1.0.0","title":"API Documentation","contact":{"name":"Site Go!","url":"https://github.com/MCC-81-Java/S4-Kelompok_02","email":"system@yopmail.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0.html"}},"host":"localhost:8080","tags":[{"name":"auth-controller","description":"Auth Controller"},{"name":"consultation-controller","description":"Consultation Controller"},{"name":"dashboard-controller","description":"Dashboard Controller"},{"name":"profile-controller","description":"Profile Controller"},{"name":"profile-status-controller","description":"Profile Status Controller"},{"name":"report-controller","description":"Report Controller"}],"paths":{"/api/v1/consultation":{"post":{"tags":["consultation-controller"],"summary":"request","operationId":"requestUsingPOST","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Consultation"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/consultation/{consultationId}":{"get":{"tags":["consultation-controller"],"summary":"getById","operationId":"getByIdUsingGET","produces":["*/*"],"parameters":[{"name":"consultationId","in":"path","description":"consultationId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Consultation"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/consultation/{consultationId}/finish":{"patch":{"tags":["consultation-controller"],"summary":"finish","operationId":"finishUsingPATCH","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"consultationId","in":"path","description":"consultationId","required":true,"type":"integer","format":"int32"},{"in":"body","name":"request","description":"request","required":true,"schema":{"$ref":"#/definitions/FinishConsultationRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Consultation"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/api/v1/consultation/{consultationId}/reject":{"patch":{"tags":["consultation-controller"],"summary":"reject","operationId":"rejectUsingPATCH","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"consultationId","in":"path","description":"consultationId","required":true,"type":"integer","format":"int32"},{"in":"body","name":"request","description":"request","required":true,"schema":{"$ref":"#/definitions/RejectConsultationRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Consultation"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/api/v1/consultation/{consultationId}/schedule":{"patch":{"tags":["consultation-controller"],"summary":"schedule","operationId":"scheduleUsingPATCH","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"consultationId","in":"path","description":"consultationId","required":true,"type":"integer","format":"int32"},{"in":"body","name":"request","description":"request","required":true,"schema":{"$ref":"#/definitions/ScheduleConsultationRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Consultation"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/api/v1/consultation/{consultationId}/validate":{"patch":{"tags":["consultation-controller"],"summary":"validate","operationId":"validateUsingPATCH","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"consultationId","in":"path","description":"consultationId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Consultation"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/api/v1/consultations":{"get":{"tags":["consultation-controller"],"summary":"getAll","operationId":"getAllUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Consultation"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/dashboard":{"get":{"tags":["dashboard-controller"],"summary":"dashboard","operationId":"dashboardUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DashboardResponse"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/login":{"post":{"tags":["auth-controller"],"summary":"login","operationId":"loginUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"request","description":"request","required":true,"schema":{"$ref":"#/definitions/LoginRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Profile"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/profile":{"put":{"tags":["profile-controller"],"summary":"update","operationId":"updateUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"request","description":"request","required":true,"schema":{"$ref":"#/definitions/UpdateProfileRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Profile"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/profile-status":{"post":{"tags":["profile-status-controller"],"summary":"create","operationId":"createUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"request","description":"request","required":true,"schema":{"$ref":"#/definitions/ProfileStatusRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ProfileStatus"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/profile-status/{statusId}":{"get":{"tags":["profile-status-controller"],"summary":"getById","operationId":"getByIdUsingGET_2","produces":["*/*"],"parameters":[{"name":"statusId","in":"path","description":"statusId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ProfileStatus"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["profile-status-controller"],"summary":"update","operationId":"updateUsingPUT_2","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"name","in":"query","required":true,"type":"string"},{"name":"statusId","in":"path","description":"statusId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ProfileStatus"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["profile-status-controller"],"summary":"update","operationId":"updateUsingDELETE","produces":["*/*"],"parameters":[{"name":"statusId","in":"path","description":"statusId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ProfileStatus"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/api/v1/profile-statuses":{"get":{"tags":["profile-status-controller"],"summary":"getAll","operationId":"getAllUsingGET_2","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/ProfileStatus"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/profile/consultations":{"get":{"tags":["consultation-controller"],"summary":"getAllByProfile","operationId":"getAllByProfileUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Consultation"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/profile/token":{"get":{"tags":["profile-controller"],"summary":"getByToken","operationId":"getByTokenUsingGET","produces":["*/*"],"parameters":[{"name":"token","in":"query","description":"token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Profile"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/profile/{profileId}":{"get":{"tags":["profile-controller"],"summary":"getById","operationId":"getByIdUsingGET_1","produces":["*/*"],"parameters":[{"name":"profileId","in":"path","description":"profileId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Profile"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["profile-controller"],"summary":"update","operationId":"updateUsingPUT_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"profileId","in":"path","description":"profileId","required":true,"type":"integer","format":"int32"},{"in":"body","name":"request","description":"request","required":true,"schema":{"$ref":"#/definitions/UpdateProfileRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Profile"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/profiles":{"get":{"tags":["profile-controller"],"summary":"getAll","operationId":"getAllUsingGET_1","produces":["*/*"],"parameters":[{"name":"roles","in":"query","description":"roles","required":true,"type":"array","items":{"type":"string"},"collectionFormat":"multi"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Profile"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/registration":{"post":{"tags":["auth-controller"],"summary":"register","operationId":"registerUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"request","description":"request","required":true,"schema":{"$ref":"#/definitions/RegistrationRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Profile"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/report/{reportId}":{"get":{"tags":["report-controller"],"summary":"getById","operationId":"getByIdUsingGET_3","produces":["*/*"],"parameters":[{"name":"reportId","in":"path","description":"reportId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Report"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["report-controller"],"summary":"update","operationId":"updateUsingPUT_3","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"reportId","in":"path","description":"reportId","required":true,"type":"integer","format":"int32"},{"in":"body","name":"request","description":"request","required":true,"schema":{"$ref":"#/definitions/WriteReportRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Report"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/v1/verification":{"post":{"tags":["auth-controller"],"summary":"verify","operationId":"verifyUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"request","description":"request","required":true,"schema":{"$ref":"#/definitions/VerificationRequest"}},{"name":"token","in":"query","description":"token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Profile"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"Consultation":{"type":"object","properties":{"consultant":{"$ref":"#/definitions/Profile"},"counselee":{"$ref":"#/definitions/Profile"},"createdAt":{"type":"string","format":"date-time"},"duration":{"type":"integer","format":"int64"},"histories":{"type":"array","items":{"$ref":"#/definitions/ConsultationHistory"}},"id":{"type":"integer","format":"int32"},"status":{"$ref":"#/definitions/ConsultationStatus"},"time":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"venue":{"type":"string"}},"title":"Consultation"},"ConsultationBar":{"type":"object","properties":{"data":{"type":"array","items":{"type":"integer","format":"int32"}},"year":{"type":"string"}},"title":"ConsultationBar"},"ConsultationCounter":{"type":"object","properties":{"numberOfConsultations":{"type":"integer","format":"int32"},"numberOfDone":{"type":"integer","format":"int32"},"numberOfRejected":{"type":"integer","format":"int32"},"numberOfScheduled":{"type":"integer","format":"int32"},"numberOfValidated":{"type":"integer","format":"int32"},"numberOfWaiting":{"type":"integer","format":"int32"}},"title":"ConsultationCounter"},"ConsultationGrowth":{"type":"object","properties":{"numberOfConsultations":{"type":"integer","format":"int32"},"numberOfConsultationsOnLastYear":{"type":"integer","format":"int32"},"numberOfConsultationsOnThisYear":{"type":"integer","format":"int32"}},"title":"ConsultationGrowth"},"ConsultationHistory":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"description":{"type":"string"},"editor":{"$ref":"#/definitions/Profile"},"id":{"$ref":"#/definitions/Id"}},"title":"ConsultationHistory"},"ConsultationStatus":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"updatedAt":{"type":"string","format":"date-time"}},"title":"ConsultationStatus"},"DashboardResponse":{"type":"object","properties":{"consultationBars":{"type":"array","items":{"$ref":"#/definitions/ConsultationBar"}},"consultationCounter":{"$ref":"#/definitions/ConsultationCounter"},"consultationGrowth":{"$ref":"#/definitions/ConsultationGrowth"},"userStat":{"$ref":"#/definitions/UserStat"}},"title":"DashboardResponse"},"FinishConsultationRequest":{"type":"object","required":["description"],"properties":{"description":{"type":"string"}},"title":"FinishConsultationRequest"},"Id":{"type":"object","properties":{"consultationStatus":{"$ref":"#/definitions/ConsultationStatus"}},"title":"Id"},"LoginRequest":{"type":"object","required":["password","username"],"properties":{"password":{"type":"string","minLength":4,"maxLength":32},"username":{"type":"string","minLength":4,"maxLength":64}},"title":"LoginRequest"},"Profile":{"type":"object","properties":{"cv":{"type":"string"},"gender":{"type":"string","enum":["FEMALE","MALE"]},"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"phone":{"type":"string"},"photo":{"type":"string"},"status":{"$ref":"#/definitions/ProfileStatus"},"user":{"$ref":"#/definitions/User"}},"title":"Profile"},"ProfileStatus":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"updatedAt":{"type":"string","format":"date-time"}},"title":"ProfileStatus"},"ProfileStatusRequest":{"type":"object","required":["name"],"properties":{"name":{"type":"string"}},"title":"ProfileStatusRequest"},"RegistrationRequest":{"type":"object","required":["email","name"],"properties":{"email":{"type":"string"},"name":{"type":"string","minLength":3,"maxLength":127}},"title":"RegistrationRequest"},"RejectConsultationRequest":{"type":"object","required":["description"],"properties":{"description":{"type":"string"}},"title":"RejectConsultationRequest"},"Report":{"type":"object","properties":{"consultation":{"$ref":"#/definitions/Consultation"},"createdAt":{"type":"string","format":"date-time"},"description":{"type":"string"},"id":{"type":"integer","format":"int32"},"updatedAt":{"type":"string","format":"date-time"}},"title":"Report"},"Role":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"updatedAt":{"type":"string","format":"date-time"}},"title":"Role"},"ScheduleConsultationRequest":{"type":"object","required":["consultantId","duration","time","venue"],"properties":{"consultantId":{"type":"integer","format":"int32"},"description":{"type":"string"},"duration":{"type":"integer","format":"int64"},"time":{"type":"integer","format":"int64"},"venue":{"type":"string"}},"title":"ScheduleConsultationRequest"},"UpdateProfileRequest":{"type":"object","required":["gender","name"],"properties":{"cv":{"type":"string"},"gender":{"type":"string","minLength":3,"maxLength":8},"name":{"type":"string","minLength":3,"maxLength":127},"phone":{"type":"string","minLength":3,"maxLength":16},"photo":{"type":"string"},"roles":{"type":"array","uniqueItems":true,"items":{"type":"string"}},"status":{"type":"string","minLength":3,"maxLength":8}},"title":"UpdateProfileRequest"},"User":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"email":{"type":"string"},"id":{"type":"integer","format":"int32"},"isEnabled":{"type":"boolean"},"password":{"type":"string"},"roles":{"type":"array","uniqueItems":true,"items":{"$ref":"#/definitions/Role"}},"token":{"type":"string"},"updatedAt":{"type":"string","format":"date-time"},"username":{"type":"string"}},"title":"User"},"UserStat":{"type":"object","properties":{"numberOfAdmins":{"type":"integer","format":"int32"},"numberOfConsultants":{"type":"integer","format":"int32"},"numberOfEmployees":{"type":"integer","format":"int32"},"numberOfUsers":{"type":"integer","format":"int32"}},"title":"UserStat"},"VerificationRequest":{"type":"object","required":["password","phone","username"],"properties":{"password":{"type":"string","minLength":4,"maxLength":32},"phone":{"type":"string","minLength":3,"maxLength":16},"username":{"type":"string","minLength":4,"maxLength":64}},"title":"VerificationRequest"},"WriteReportRequest":{"type":"object","required":["description"],"properties":{"description":{"type":"string"}},"title":"WriteReportRequest"}}}