# serverapp/src/main/java/id/co/mii/sitego/model/Report.java
package id.co.mii.sitego.model;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import javax.persistence.*;
import java.time.LocalDateTime;

@Entity
@Table(name = "report")
@NoArgsConstructor
@AllArgsConstructor
@Setter
@Getter
public class Report {
    @Id
    @Column(name = "id")
    private Integer id;

    @Column(name = "description", nullable = false)
    @Lob
    private String description;

    @Column(name = "updated_at", nullable = false)
    private LocalDateTime updatedAt = LocalDateTime.now();

    @Column(name = "created_at", nullable = false)
    private LocalDateTime createdAt = LocalDateTime.now();

    @OneToOne
    @MapsId
    @JoinColumn(name = "id")
    private Consultation consultation;
}