# serverapp/src/main/java/id/co/mii/sitego/model/request/FinishConsultationRequest.java
package id.co.mii.sitego.model.request;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import javax.validation.constraints.NotBlank;

@NoArgsConstructor
@AllArgsConstructor
@Data
public class FinishConsultationRequest {
    @NotBlank
    private String description;
}