# serverapp/src/main/java/id/co/mii/sitego/model/request/RejectConsultationRequest.java
package id.co.mii.sitego.model.request;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import javax.validation.constraints.NotBlank;

@NoArgsConstructor
@AllArgsConstructor
@Data
public class RejectConsultationRequest {
    @NotBlank
    private String description;
}